Use chooser viewset to construct document chooser widget
